Python/FastAPI/Project Blog/requirements.txt
fastapi
uvicorn[standard]

sqlmodel
sqlalchemy

psycopg[binary]

passlib[bcrypt]
bcrypt

python-jose[cryptography]
python-multipart

pydantic

